+++ _docs/aleph.config.ts
@@ -1,3 +1,4 @@
+// Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
 import markdown from "https://deno.land/x/aleph/plugins/markdown.ts";
 import type { Config } from "aleph/types";
 import windicss from "./plugins/windicss.ts";