+++ test/mod.ts
@@ -1,7 +1,9 @@
 // Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
 import { AnyFn } from "../_types.ts";
-import { isString } from "../deps.ts";
+import { AssertionError, fromFileUrl, isString } from "../deps.ts";
 import { each } from "./each.ts";
+import { parseStackTrace } from "../report/stack_trace.ts";
+import { prettyCode } from "../report/assertion.ts";
 
 type Test<Context extends Record<PropertyKey, unknown>> = {
   setup?: () => Context | void;
@@ -36,11 +38,42 @@ function _test<T extends Record<PropertyKey, unknown>>(
     const { setup, teardown, fn, ...rest } = t;
     Deno.test({
       ...rest,
-      fn: (denoContext) => {
+      fn: async (denoContext) => {
         const context = setup?.() ?? {} as T;
 
         try {
-          fn({ ...context, ...denoContext });
+          await fn({ ...context, ...denoContext });
+        } catch (e) {
+          if (e instanceof Error) {
+            if (e.stack) {
+              const { state } = await Deno.permissions.query({
+                name: "read",
+              });
+
+              if (state !== "granted") {
+                throw e;
+              }
+
+              const stackTraces = parseStackTrace(e.stack);
+              const testFileStack = stackTraces.find(({ file }) => {
+                return /file:\/\/\S+_test\.[j|t]sx?$/.test(file);
+              });
+
+              if (!testFileStack) {
+                throw e;
+              }
+
+              const { lineNumber, column, file } = testFileStack;
+              const code = await Deno.readTextFile(
+                fromFileUrl(file),
+              );
+
+              const pretty = prettyCode({ code, lineNumber, column });
+              throw new AssertionError(`${e.message}\n${pretty}\n`);
+            }
+          } else {
+            throw Error("panic: unexpected error");
+          }
         } finally {
           teardown?.(context);
         }